backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.8.0
reportlab==4.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.2.0
